CBFS: use cbfs_get_file_content whenever possible rather than cbfs_get_file
Number one reason to use cbfs_get_file was to get file length. With prvious
patch no more need for this.

+++ src/vendorcode/google/chromeos/chromeos.h
@@ -48,7 +48,7 @@ void init_chromeos(int bootmode);
 #if CONFIG_VBOOT_VERIFY_FIRMWARE
 struct romstage_handoff;
 void vboot_verify_firmware(struct romstage_handoff *handoff);
-void *vboot_get_payload(int *len);
+void *vboot_get_payload(size_t *len);
 /* Returns 0 on success < 0 on error. */
 int vboot_get_handoff_info(void **addr, uint32_t *size);
 #endif
